Refactor: Split monolithic main.rs into multiple modules
- I created a library structure (src/lib.rs) and broke down the application logic into several modules: cli, config, ftp_ops, instance, logging, and shutdown.
- I relocated relevant functions, structures, and constants to their appropriate modules to improve code organization and maintainability.
- I refactored src/main.rs to act as a thin wrapper, utilizing the newly created library.
- I addressed and resolved various issues that emerged during the refactoring process, including: incorrect import paths, type mismatches due to differing versions of the 'chrono' crate between the main project and its dependencies, and errors in SystemTime conversion logic.
- I updated the Makefile to ensure 'cargo build' is executed before 'test.sh' to make the binary available for end-to-end tests.
- I installed 'pyftpdlib' dependency for end-to-end tests.
- I ensured that all tests, including unit tests (`cargo test`) and end-to-end tests (`make test`), pass successfully with the new modular structure.
